fixed last 409 conflicts + compilation schema rollback

+++ core/category-service/src/main/java/ru/yandex/practicum/service/impls/AdminCategoryServiceImpl.java
@@ -8,7 +8,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.yandex.practicum.category.dtos.CategoryDto;
 import ru.yandex.practicum.category.dtos.NewCategoryDto;
+import ru.yandex.practicum.errors.exceptions.ConflictException;
 import ru.yandex.practicum.errors.exceptions.NotFoundException;
+import ru.yandex.practicum.event.PublicEventFeignClient;
 import ru.yandex.practicum.mapper.CategoryMapper;
 import ru.yandex.practicum.model.Category;
 import ru.yandex.practicum.repository.CategoryRepository;
@@ -22,10 +24,14 @@ import ru.yandex.practicum.service.AdminCategoryService;
 public class AdminCategoryServiceImpl implements AdminCategoryService {
     CategoryMapper mapper;
     CategoryRepository repository;
+    PublicEventFeignClient eventClient;
 
     @Override
     public CategoryDto createNewCategory(NewCategoryDto categoryDto) {
         log.info("Запрос в сервис на создание новой категории");
+        if (repository.existsByName(categoryDto.getName())) {
+            throw new ConflictException("Такое название категории уже существует");
+        }
         Category category = repository.save(mapper.toCategoryEntity(categoryDto));
         log.debug("Категория с id = {} создана", category.getId());
         return mapper.toCategoryDto(category);
@@ -34,6 +40,9 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     @Override
     public void deleteCategory(Long categoryId) {
         log.info("Запрос в сервис на удаление категории");
+        if (eventClient.checkEventsByCategoryId(categoryId)) {
+            throw new ConflictException("Категория в данный момент используется событиями");
+        }
         repository.deleteById(categoryId);
         log.info("Запрос на удаление успешно выполнен");
     }
@@ -43,6 +52,9 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
         log.info("Запрос в сервис на обновление категории");
         Category category = repository.findById(categoryId)
                 .orElseThrow(() -> new NotFoundException("Категория с id " + categoryId + " не найдена"));
+        if (repository.existsByName(categoryDto.getName())) {
+            throw new ConflictException("Такое название категории уже существует");
+        }
         mapper.updateCategoryFromDto(categoryDto, category);
         category = repository.save(category);
         log.debug("Категория с id = {} успешно обновлена", category.getId());
